testplans: Add build proto and repo tool path to test plan request
The serialized build proto field removes the need for writing >100 build
protos to disk (creating a bunch of ugly write_build_ steps in Milo).

The repo tool path is needed because the test planner binary invokes the
repo tool and I'd previously assumed that repo would be on the PATH of
the bots. That wasn't true :/.

BUG=chromium:922995
TEST=generate.sh

Change-Id: Ia5a864448da6ccf627170b9ce798dc8fcb8c3792
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/proto/+/1565584
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>