Run generate.sh script in proto repo
All of these "device" pb.go files resulted from running generate.sh. They're
not my protos, but this keeps the generated code in sync with the src
protos.

Yes, we should probably have a presubmit that checks that the generated
code is in sync with the src protos.

Change-Id: I19b06e8b6ee29c5a07beffd7e6af8acd4aeab0fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/proto/+/1565585
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>