testplans: Add skylab_board field to be used for scheduling tests
This is what we'll use to tell Skylab which board it should
use for testing. This is distinct from a build's build target, since
sometimes the two aren't equal. e.g. caroline-arcnext builds need to be
tested on caroline devices.

This concept is irrelevant at this point for VM tests.

BUG=chromium:954515
TEST=generated protos

Change-Id: I4f75f8463bcce29ef6a3eeeb2cbb849fc9d712fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/proto/+/1575431
Tested-by: Sean Abraham <seanabraham@chromium.org>
Reviewed-by: Evan Hernandez <evanhernandez@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>