infra/proto: Add OWNERS file
Enforcement is not enabled yet. This CL is a suggestion of ownership
candidates based on git activity and you are welcome to add other owners
as you see fit. See the following link for more information on the syntax
for OWNERS files in ChromiumOS repositories:
https://gerrit.googlesource.com/plugins/find-owners/+/master/src/main/resources/Documentation/syntax.md

BUG=b:132095048
TEST=None

Change-Id: I730df6f70a48e6b37b5203906703b7b7c12dea28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/proto/+/1599382
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>